// scripts/Pix-Scripting/src/types.ts
export interface PixivImageUrls {
  square_medium?: string
  medium?: string
  large?: string
  original?: string
}

export interface PixivTag {
  name: string
  translated_name?: string | null
}

export interface PixivUser {
  id: number
  name: string
  account: string
  profile_image_urls?: { medium?: string }
  comment?: string
  is_followed?: boolean
  is_mypixiv?: boolean
}

export interface PixivSeries {
  id: number
  title?: string | null
}

export interface PixivWatchlistSeries {
  id: number
  title: string
  url?: string | null
  mask_text?: string | null
  published_content_count: number
  last_published_content_datetime?: string | null
  latest_content_id?: number | null
  user?: PixivUser | null
}

export interface PixivMetaPage {
  image_urls: PixivImageUrls
}

export interface PixivMetaSinglePage {
  original_image_url?: string
}

export interface PixivIllustration {
  id: number
  title: string
  type: "illust" | "manga" | "ugoira"
  image_urls: PixivImageUrls
  caption: string
  user: PixivUser
  tags: PixivTag[]
  create_date: string
  page_count: number
  width: number
  height: number
  x_restrict: number
  series?: PixivSeries | null
  episode_number?: number
  meta_single_page?: PixivMetaSinglePage
  meta_pages: PixivMetaPage[]
  total_view: number
  total_bookmarks: number
  is_bookmarked: boolean
  is_muted: boolean
  illust_ai_type: number
  total_comments: number
  comment_access_control: number
}
// ...
export interface PixivUserDetail {
  user: PixivUser
  profile: {
    webpage?: string
    gender?: string
    birth?: string
    region?: string
    job?: string
    total_follow_users: number
    total_mypixiv_users: number
    total_illusts: number
    total_manga: number
    total_novels: number
    total_illust_bookmarks_public: number
    total_illust_series: number
    background_image_url?: string
    twitter_account?: string
    twitter_url?: string
    is_premium: boolean
  }
  workspace?: {
    pc?: string
    monitor?: string
    tool?: string
    tablet?: string
    music?: string
    desk?: string
    chair?: string
    comment?: string
  } | null
}
// ...
export interface PixivNovel {
  id: number
  title: string
  caption?: string
  user: PixivUser
  tags: PixivTag[]
  create_date: string
  page_count: number
  x_restrict: number
  total_view: number
  total_bookmarks: number
  is_bookmarked: boolean
  is_muted: boolean
  novel_ai_type?: number
  total_comments: number
  text_length?: number
  visible: boolean
  series?: PixivSeries | null
  episode_number?: number
  novel_marker?: {
    page: number
  } | null
  image_urls?: {
    square_medium?: string
    medium?: string
    large?: string
  } | null
  cover?: {
    urls?: { "240mw"?: string; "480mw"?: string; "1200x1200"?: string }
  } | null
}